Escargot: refactor.
